Rename SingleGroup to GroupDetail, SingleExpense to ExpenseDetail
